floats are not ints
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_0branch/bzflag@13199 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/GuidedMissleStrategy.cxx
@@ -430,8 +430,8 @@ void GuidedMissileStrategy::expire()
 void GuidedMissileStrategy::radarRender() const
 {
   const float *orig = getPath().getPosition();
-  const int length = BZDBCache::linedRadarShots;
-  const int size   = BZDBCache::sizedRadarShots;
+  const int length = (int)BZDBCache::linedRadarShots;
+  const int size   = (int)BZDBCache::sizedRadarShots;
 
   float shotTailLength = BZDB.eval(StateDatabase::BZDB_SHOTTAILLENGTH);
   // Display leading lines
